requirements.txt
absl-py==1.4.0
asttokens==2.2.1
astunparse==1.6.3
attrs==22.2.0
backcall==0.2.0
cachetools==5.3.0
certifi==2023.5.7
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.1.0
colorama==0.4.6
comm==0.1.3
contourpy==1.2.0
cryptography==44.0.0
cycler==0.12.1
data==0.4
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
download==0.3.5
executing==1.2.0
flatbuffers==23.5.9
fonttools==4.45.0
funcsigs==1.0.2
future==1.0.0
gast==0.4.0
google-auth==2.18.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.55.0
h5py==3.8.0
idna==3.4
ipykernel==6.22.0
ipython==8.12.0
jedi==0.18.2
jupyter_client==8.1.0
jupyter_core==5.3.0
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
latex==0.7.0
libclang==16.0.0
lxml==5.2.2
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.8.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.24.2
oauthlib==3.2.2
opencv-python==4.5.5.62
opt-einsum==3.3.0
packaging==23.0
pandas==2.0.0
parso==0.8.3
pdf2image==1.17.0
pdf2txt==0.7.1
pdfminer==20191125
pdfminer.six==20200726
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.2.0
prompt-toolkit==3.0.38
protobuf==3.19.6
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.22
pycryptodome==3.19.0
Pygments==2.14.0
PyMuPDF==1.24.2
PyMuPDFb==1.24.1
pyparsing==3.1.1
PyPDF2==3.0.1
pyserial==3.5
python-dateutil==2.8.2
python-pptx==0.6.23
pytz==2023.3
pywin32==306
pyxml2pdf==0.3.4
pyzmq==25.0.2
reportlab==4.2.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
shutilwhich==1.1.0
six==1.16.0
sortedcontainers==2.4.0
stack-data==0.6.2
tempdir==0.7.1
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.1
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
tornado==6.2
tqdm==4.66.4
traitlets==5.9.0
typing_extensions==4.6.0
tzdata==2023.3
urllib3==1.26.16
wcwidth==0.2.6
Werkzeug==2.3.4
wrapt==1.15.0
XlsxWriter==3.2.0
